+++ tht-apis/src/test/java/com/tht/thtapis/documentation/DailyFallingDocumentation.java
@@ -2,6 +2,7 @@ package com.tht.thtapis.documentation;
 
 import com.epages.restdocs.apispec.ResourceSnippetParameters;
 import com.epages.restdocs.apispec.Schema;
+import com.tht.enums.EnumDocsUtils;
 import com.tht.enums.dailyfalling.DailyFallingType;
 import com.tht.thtapis.controller.config.ControllerTestConfig;
 import com.tht.thtapis.controller.config.WithCustomMockUser;
