+++ tht-apis/src/test/java/com/tht/thtapis/documentation/UserJoinDocumentation.java
@@ -2,6 +2,7 @@ package com.tht.thtapis.documentation;
 
 import com.epages.restdocs.apispec.ResourceSnippetParameters;
 import com.epages.restdocs.apispec.Schema;
+import com.tht.enums.EnumDocsUtils;
 import com.tht.enums.agreement.AgreementCategory;
 import com.tht.enums.user.UserFrequency;
 import com.tht.enums.user.UserReligion;
